feet : 글 상세조회 및 댓글 에러 수정
코드 병합 과정에서 일어난 에러 수정

+++ src/main/java/com/lol/community/comment/dto/CommentPageRequestDTO.java
@@ -12,7 +12,7 @@ import org.springframework.data.domain.Sort;
 @AllArgsConstructor
 public class CommentPageRequestDTO {
 
-  private Long board_id;
+  private Integer board_id;
   private int page;
   private int size;
