+++ src/main/java/com/lol/community/user/repository/UserRepository.java
@@ -3,6 +3,5 @@ package com.lol.community.user.repository;
 import com.lol.community.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
